Upadates

+++ app/Http/Controllers/UploadFilesController.php
@@ -8,17 +8,21 @@ use Illuminate\Support\Facades\DB;
 
 class UploadFilesController extends Controller
 {
+    public function getAll(){
+      return response()->json(UploadFiles::all());
+    }
+
     public function getFiles($page = 1, $keyword = null){
-      $paginate = ($page-1) * 5;
+      $paginate = ($page-1) * 6;
       if($keyword === null) {
         $files = DB::table('upload_files')
-                    ->skip($paginate)->take(5)->whereNull("deleted_at")
+                    ->skip($paginate)->take(6)->whereNull("deleted_at")
                     ->orderBy('created_at', 'desc')->get();
       }else{
         $files = DB::table('upload_files')
                 ->where("name", "LIKE","%$keyword%")
                 ->orWhere("filesrc", "LIKE", "%$keyword%")
-                ->skip($paginate)->take(5)->whereNull("deleted_at")
+                ->skip($paginate)->take(6)->whereNull("deleted_at")
                 ->orderBy('created_at', 'desc')->get();
       }
 
@@ -53,17 +57,21 @@ class UploadFilesController extends Controller
 
     public function deleteFile($id)
     {
-        UploadFiles::findOrFail($id)->delete();
+        $file = UploadFiles::find($id);
+        if ($file !== null) {
+          unlink( base_path('public')."/documents/".$file->filesrc );
+          UploadFiles::findOrFail($id)->delete();
+        }
         return response('Deleted Successfully', 200);
     }
 
     public function showHistory($page = null)
     {
       if ($page !== null) {
-        $paginate = ($page-1) * 5;
-        $files = DB::table('upload_files')->skip($paginate)->take(5)->get();
+        $paginate = ($page-1) * 6;
+        $files = DB::table('upload_files')->skip($paginate)->take(6)->get();
       }else{
-        $files = DB::table('upload_files')->take(5)->get();
+        $files = DB::table('upload_files')->take(6)->get();
       }
       return response()->json($files);
     }
